13. przeorganizowanie projektu zmieniejac klase Node z publicznej na prywatna, wewnatrz klasy IntLinkedList
