+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -2,6 +2,41 @@ package edu.kis.vh.nursery.list;
 
 public class  IntLinkedList {
 
+	private static class Node {
+
+		private int value;
+		private Node prev;
+		private Node next;
+
+		public Node(int i) {
+			setValue(i);
+		}
+
+		public int getValue() {
+			return value;
+		}
+
+		public void setValue(int value) {
+			this.value = value;
+		}
+
+		public Node getPrev() {
+			return prev;
+		}
+
+		public void setPrev(Node prev) {
+			this.prev = prev;
+		}
+
+		public Node getNext() {
+			return next;
+		}
+
+		public void setNext(Node next) {
+			this.next = next;
+		}
+	}
+
 	private Node last;
 	private int i;
